requirements.txt
PySide6>=6.5
keyring
cryptography
PyYAML
reportlab
pypdf
requests